Moved temporary_buffer_file to planscore.util module
