+++ planscore/score.py
@@ -15,24 +15,12 @@ def score_plan(upload, plan_path, tiles_prefix):
         raise RuntimeError('Could not open file')
     
     for (index, feature) in enumerate(ds.GetLayer(0)):
-        upload.tiles.append([])
         feature_count += 1
-        geometry = feature.GetGeometryRef()
-        geometry.TransformTo(prepare_state.EPSG4326)
-        
-        xxyy_extent = geometry.GetEnvelope()
-        tiles = prepare_state.iter_extent_tiles(xxyy_extent, prepare_state.TILE_ZOOM)
         print(index, feature, file=output)
-        for (coord, tile_wkt) in tiles:
-            tile_zxy = '{zoom}/{column}/{row}'.format(**coord.__dict__)
-            tile_geom = ogr.CreateGeometryFromWkt(tile_wkt)
-            
-            if not tile_geom.Intersects(geometry):
-                continue
-            
-            upload.tiles[-1].append(tile_zxy)
-            print(' ', prepare_state.KEY_FORMAT.format(state='XX',
-                zxy=tile_zxy), file=output)
+
+        tiles, district_output = score_district(feature.GetGeometryRef(), tiles_prefix)
+        upload.tiles.append(tiles)
+        output.write(district_output)
     
     length = os.stat(plan_path).st_size
     
@@ -40,3 +28,24 @@ def score_plan(upload, plan_path, tiles_prefix):
         length, os.path.basename(plan_path)), file=output) 
     
     return output.getvalue()
+
+def score_district(geometry, tiles_prefix):
+    '''
+    '''
+    tile_list, output = [], io.StringIO()
+    geometry.TransformTo(prepare_state.EPSG4326)
+    
+    xxyy_extent = geometry.GetEnvelope()
+    tiles = prepare_state.iter_extent_tiles(xxyy_extent, prepare_state.TILE_ZOOM)
+    for (coord, tile_wkt) in tiles:
+        tile_zxy = '{zoom}/{column}/{row}'.format(**coord.__dict__)
+        tile_geom = ogr.CreateGeometryFromWkt(tile_wkt)
+        
+        if not tile_geom.Intersects(geometry):
+            continue
+        
+        tile_list.append(tile_zxy)
+        print(' ', prepare_state.KEY_FORMAT.format(state='XX',
+            zxy=tile_zxy), file=output)
+    
+    return tile_list, output.getvalue()
